+++ src/Humanizer/Configuration/NumberToWordsConverterRegistry.cs
@@ -4,7 +4,8 @@ namespace Humanizer.Configuration
 {
     internal class NumberToWordsConverterRegistry : LocaliserRegistry<INumberToWordsConverter>
     {
-        public NumberToWordsConverterRegistry() : base((culture) => new DefaultNumberToWordsConverter(culture))
+        public NumberToWordsConverterRegistry()
+            : base((culture) => new DefaultNumberToWordsConverter(culture))
         {
             Register("en", new EnglishNumberToWordsConverter());
             Register("ar", new ArabicNumberToWordsConverter());
@@ -21,8 +22,11 @@ namespace Humanizer.Configuration
             Register("bn-BD", new BanglaNumberToWordsConverter());
             Register("tr", new TurkishNumberToWordConverter());
             Register("it", new ItalianNumberToWordsConverter());
+            Register("uk", new UkrainianNumberToWordsConverter());
             Register("uz-Latn-UZ", new UzbekLatnNumberToWordConverter());
             Register("uz-Cyrl-UZ", new UzbekCyrlNumberToWordConverter());
+            Register("sr", (culture) => new SerbianCyrlNumberToWordsConverter(culture));
+            Register("sr-Latn", (culture) => new SerbianNumberToWordsConverter(culture));
         }
     }
 }