+++ src/Humanizer/Configuration/OrdinalizerRegistry.cs
@@ -10,9 +10,11 @@ namespace Humanizer.Configuration
             Register("en", new EnglishOrdinalizer());
             Register("es", new SpanishOrdinalizer());
             Register("it", new ItalianOrdinalizer());
-            Register("pt-BR", new BrazilianPortugueseOrdinalizer());
+            Register("nl", new DutchOrdinalizer());
+            Register("pt", new PortugueseOrdinalizer());
             Register("ru", new RussianOrdinalizer());
             Register("tr", new TurkishOrdinalizer());
+            Register("uk", new UkrainianOrdinalizer());
         }
     }
 }